refactor: rename tb_embeds to tb_ui and reduce intent requirements
