+++ tb_discord/tb_commands/roles.py
@@ -1,49 +1,11 @@
 from datetime import datetime, timedelta
-from discord.ui import Button, RoleSelect, View
-from discord import app_commands, ButtonStyle, Message, Interaction, Role, TextChannel, utils
+from discord import app_commands, Message, Interaction, TextChannel, utils
 from tb_db import sql_op
-from tb_discord.tb_embeds.public import RoleButtonEmbed
+from tb_discord.tb_ui import RoleButtonEmbed, RolesView, RoleChoiceView
 from tb_discord.data_structures import RolesMessage, role_messages
 
 
-__all__ = ['command_list', 'RolesView']
-
-
-
-class RoleChoiceView(View):
-	def __init__(self, message: Message):
-		super().__init__(timeout=120)
-		self.message = message
-		self.roles = TBRolesSelect()
-		self.add_item(self.roles)
-
-
-class TBRolesSelect(RoleSelect):
-	def __init__(self):
-		super().__init__(min_values=1, max_values=10)
-
-	async def callback(self, interaction: Interaction):
-		self.view.stop()
-
-
-class RolesView(View):
-	def __init__(self, roles: list[Role]):
-		super().__init__(timeout=None)
-		for role in roles:
-			self.add_item(RoleButton(role))
-
-
-class RoleButton(Button):
-	def __init__(self, role: Role):
-		super().__init__(style=ButtonStyle.primary, label=role.name)
-		self.role = role
-
-	async def callback(self, interaction: Interaction):
-		if self.role in interaction.user.roles:
-			await interaction.user.remove_roles(self.role)
-		else:
-			await interaction.user.add_roles(self.role)
-		await interaction.response.defer()
+__all__ = ['command_list']
 
 
 @app_commands.command()
