+++ tb_discord/tb_ui/__init__.py
@@ -1,0 +1,3 @@
+"""Discord embed subclasses for self-containted updates"""
+from tb_discord.tb_ui.private import *
+from tb_discord.tb_ui.public import *
